pyproject.toml
[tool.poetry]
name = "web-processing"
version = "0.1.0"
description = ""
authors = ["Mohammad Saad <81032267+masaad01@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.31.0"
python-dotenv = "^1.0.1"
fastapi = "^0.110.3"
uvicorn = "^0.29.0"
speedtest-cli = "^2.1.3"
country-converter = "^1.2"
apihole = "^0.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
